[issue#0007] Split code into multiple pairs of each peripherals.
- refer to below.
https://community.st.com/t5/stm32cubemx-mcus/is-it-possible-to-automatically-separate-the-code-generated-by/td-p/71174
