+++ Project/Alm/Core/Src/main.c
@@ -19,6 +19,9 @@
 /* Includes ------------------------------------------------------------------*/
 #include "main.h"
 #include "cmsis_os.h"
+#include "tim.h"
+#include "usart.h"
+#include "gpio.h"
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
@@ -43,67 +46,14 @@
 /* USER CODE END PM */
 
 /* Private variables ---------------------------------------------------------*/
-TIM_HandleTypeDef htim2;
 
-UART_HandleTypeDef huart2;
-
-/* Definitions for defaultTask */
-osThreadId_t defaultTaskHandle;
-const osThreadAttr_t defaultTask_attributes = {
-  .name = "defaultTask",
-  .priority = (osPriority_t) osPriorityNormal,
-  .stack_size = 128 * 4
-};
-/* Definitions for TSK_LED */
-osThreadId_t TSK_LEDHandle;
-const osThreadAttr_t TSK_LED_attributes = {
-  .name = "TSK_LED",
-  .priority = (osPriority_t) osPriorityLow,
-  .stack_size = 256 * 4
-};
-/* Definitions for TSK_BTN */
-osThreadId_t TSK_BTNHandle;
-const osThreadAttr_t TSK_BTN_attributes = {
-  .name = "TSK_BTN",
-  .priority = (osPriority_t) osPriorityLow,
-  .stack_size = 256 * 4
-};
-/* Definitions for TSK_SWD */
-osThreadId_t TSK_SWDHandle;
-const osThreadAttr_t TSK_SWD_attributes = {
-  .name = "TSK_SWD",
-  .priority = (osPriority_t) osPriorityLow,
-  .stack_size = 256 * 4
-};
-/* Definitions for queLedTask */
-osMessageQueueId_t queLedTaskHandle;
-const osMessageQueueAttr_t queLedTask_attributes = {
-  .name = "queLedTask"
-};
-/* Definitions for queBtnTask */
-osMessageQueueId_t queBtnTaskHandle;
-const osMessageQueueAttr_t queBtnTask_attributes = {
-  .name = "queBtnTask"
-};
-/* Definitions for queSwdTask */
-osMessageQueueId_t queSwdTaskHandle;
-const osMessageQueueAttr_t queSwdTask_attributes = {
-  .name = "queSwdTask"
-};
 /* USER CODE BEGIN PV */
 
 /* USER CODE END PV */
 
 /* Private function prototypes -----------------------------------------------*/
 void SystemClock_Config(void);
-static void MX_GPIO_Init(void);
-static void MX_USART2_UART_Init(void);
-static void MX_TIM2_Init(void);
-void StartDefaultTask(void *argument);
-void vidTaskLed(void *argument);
-void vidTaskButton(void *argument);
-void vidTaskSwd(void *argument);
-
+void MX_FREERTOS_Init(void);
 /* USER CODE BEGIN PFP */
 
 /* USER CODE END PFP */
@@ -151,52 +101,8 @@ int main(void)
   /* Init scheduler */
   osKernelInitialize();
 
-  /* USER CODE BEGIN RTOS_MUTEX */
-  /* add mutexes, ... */
-  /* USER CODE END RTOS_MUTEX */
-
-  /* USER CODE BEGIN RTOS_SEMAPHORES */
-  /* add semaphores, ... */
-  /* USER CODE END RTOS_SEMAPHORES */
-
-  /* USER CODE BEGIN RTOS_TIMERS */
-  /* start timers, add new ones, ... */
-  /* USER CODE END RTOS_TIMERS */
-
-  /* Create the queue(s) */
-  /* creation of queLedTask */
-  queLedTaskHandle = osMessageQueueNew (20, sizeof(uint8_t), &queLedTask_attributes);
-
-  /* creation of queBtnTask */
-  queBtnTaskHandle = osMessageQueueNew (20, sizeof(uint8_t), &queBtnTask_attributes);
-
-  /* creation of queSwdTask */
-  queSwdTaskHandle = osMessageQueueNew (20, sizeof(uint8_t), &queSwdTask_attributes);
-
-  /* USER CODE BEGIN RTOS_QUEUES */
-  /* add queues, ... */
-  /* USER CODE END RTOS_QUEUES */
-
-  /* Create the thread(s) */
-  /* creation of defaultTask */
-  defaultTaskHandle = osThreadNew(StartDefaultTask, NULL, &defaultTask_attributes);
-
-  /* creation of TSK_LED */
-  TSK_LEDHandle = osThreadNew(vidTaskLed, NULL, &TSK_LED_attributes);
-
-  /* creation of TSK_BTN */
-  TSK_BTNHandle = osThreadNew(vidTaskButton, NULL, &TSK_BTN_attributes);
-
-  /* creation of TSK_SWD */
-  TSK_SWDHandle = osThreadNew(vidTaskSwd, NULL, &TSK_SWD_attributes);
-
-  /* USER CODE BEGIN RTOS_THREADS */
-  /* add threads, ... */
-  /* USER CODE END RTOS_THREADS */
-
-  /* USER CODE BEGIN RTOS_EVENTS */
-  /* add events, ... */
-  /* USER CODE END RTOS_EVENTS */
+  /* Call init function for freertos objects (in cmsis_os2.c) */
+  MX_FREERTOS_Init();
 
   /* Start scheduler */
   osKernelStart();
@@ -259,217 +165,9 @@ void SystemClock_Config(void)
   }
 }
 
-/**
-  * @brief TIM2 Initialization Function
-  * @param None
-  * @retval None
-  */
-static void MX_TIM2_Init(void)
-{
-
-  /* USER CODE BEGIN TIM2_Init 0 */
-
-  /* USER CODE END TIM2_Init 0 */
-
-  TIM_ClockConfigTypeDef sClockSourceConfig = {0};
-  TIM_MasterConfigTypeDef sMasterConfig = {0};
-  TIM_OC_InitTypeDef sConfigOC = {0};
-
-  /* USER CODE BEGIN TIM2_Init 1 */
-
-  /* USER CODE END TIM2_Init 1 */
-  htim2.Instance = TIM2;
-  htim2.Init.Prescaler = 0;
-  htim2.Init.CounterMode = TIM_COUNTERMODE_UP;
-  htim2.Init.Period = 1599;
-  htim2.Init.ClockDivision = TIM_CLOCKDIVISION_DIV1;
-  htim2.Init.AutoReloadPreload = TIM_AUTORELOAD_PRELOAD_DISABLE;
-  if (HAL_TIM_Base_Init(&htim2) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  sClockSourceConfig.ClockSource = TIM_CLOCKSOURCE_INTERNAL;
-  if (HAL_TIM_ConfigClockSource(&htim2, &sClockSourceConfig) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  if (HAL_TIM_PWM_Init(&htim2) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  sMasterConfig.MasterOutputTrigger = TIM_TRGO_RESET;
-  sMasterConfig.MasterSlaveMode = TIM_MASTERSLAVEMODE_DISABLE;
-  if (HAL_TIMEx_MasterConfigSynchronization(&htim2, &sMasterConfig) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  sConfigOC.OCMode = TIM_OCMODE_PWM1;
-  sConfigOC.Pulse = 0;
-  sConfigOC.OCPolarity = TIM_OCPOLARITY_HIGH;
-  sConfigOC.OCFastMode = TIM_OCFAST_DISABLE;
-  if (HAL_TIM_PWM_ConfigChannel(&htim2, &sConfigOC, TIM_CHANNEL_1) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  /* USER CODE BEGIN TIM2_Init 2 */
-
-  /* USER CODE END TIM2_Init 2 */
-  HAL_TIM_MspPostInit(&htim2);
-
-}
-
-/**
-  * @brief USART2 Initialization Function
-  * @param None
-  * @retval None
-  */
-static void MX_USART2_UART_Init(void)
-{
-
-  /* USER CODE BEGIN USART2_Init 0 */
-
-  /* USER CODE END USART2_Init 0 */
-
-  /* USER CODE BEGIN USART2_Init 1 */
-
-  /* USER CODE END USART2_Init 1 */
-  huart2.Instance = USART2;
-  huart2.Init.BaudRate = 115200;
-  huart2.Init.WordLength = UART_WORDLENGTH_8B;
-  huart2.Init.StopBits = UART_STOPBITS_1;
-  huart2.Init.Parity = UART_PARITY_NONE;
-  huart2.Init.Mode = UART_MODE_TX_RX;
-  huart2.Init.HwFlowCtl = UART_HWCONTROL_NONE;
-  huart2.Init.OverSampling = UART_OVERSAMPLING_16;
-  huart2.Init.OneBitSampling = UART_ONE_BIT_SAMPLE_DISABLE;
-  huart2.Init.ClockPrescaler = UART_PRESCALER_DIV1;
-  huart2.AdvancedInit.AdvFeatureInit = UART_ADVFEATURE_NO_INIT;
-  if (HAL_UART_Init(&huart2) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  if (HAL_UARTEx_SetTxFifoThreshold(&huart2, UART_TXFIFO_THRESHOLD_1_8) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  if (HAL_UARTEx_SetRxFifoThreshold(&huart2, UART_RXFIFO_THRESHOLD_1_8) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  if (HAL_UARTEx_DisableFifoMode(&huart2) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  /* USER CODE BEGIN USART2_Init 2 */
-
-  /* USER CODE END USART2_Init 2 */
-
-}
-
-/**
-  * @brief GPIO Initialization Function
-  * @param None
-  * @retval None
-  */
-static void MX_GPIO_Init(void)
-{
-  GPIO_InitTypeDef GPIO_InitStruct = {0};
-/* USER CODE BEGIN MX_GPIO_Init_1 */
-/* USER CODE END MX_GPIO_Init_1 */
-
-  /* GPIO Ports Clock Enable */
-  __HAL_RCC_GPIOC_CLK_ENABLE();
-  __HAL_RCC_GPIOF_CLK_ENABLE();
-  __HAL_RCC_GPIOA_CLK_ENABLE();
-
-  /*Configure GPIO pin : B1_Pin */
-  GPIO_InitStruct.Pin = B1_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_IT_RISING;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  HAL_GPIO_Init(B1_GPIO_Port, &GPIO_InitStruct);
-
-/* USER CODE BEGIN MX_GPIO_Init_2 */
-/* USER CODE END MX_GPIO_Init_2 */
-}
-
 /* USER CODE BEGIN 4 */
 
 /* USER CODE END 4 */
-
-/* USER CODE BEGIN Header_StartDefaultTask */
-/**
-  * @brief  Function implementing the defaultTask thread.
-  * @param  argument: Not used
-  * @retval None
-  */
-/* USER CODE END Header_StartDefaultTask */
-void StartDefaultTask(void *argument)
-{
-  /* USER CODE BEGIN 5 */
-  /* Infinite loop */
-  for(;;)
-  {
-    osDelay(1);
-  }
-  /* USER CODE END 5 */
-}
-
-/* USER CODE BEGIN Header_vidTaskLed */
-/**
-* @brief Function implementing the TSK_LED thread.
-* @param argument: Not used
-* @retval None
-*/
-/* USER CODE END Header_vidTaskLed */
-void vidTaskLed(void *argument)
-{
-  /* USER CODE BEGIN vidTaskLed */
-  /* Infinite loop */
-  for(;;)
-  {
-    vidTaskLedProcess();
-    osDelay(1);
-  }
-  /* USER CODE END vidTaskLed */
-}
-
-/* USER CODE BEGIN Header_vidTaskButton */
-/**
-* @brief Function implementing the TSK_BTN thread.
-* @param argument: Not used
-* @retval None
-*/
-/* USER CODE END Header_vidTaskButton */
-void vidTaskButton(void *argument)
-{
-  /* USER CODE BEGIN vidTaskButton */
-  /* Infinite loop */
-  for(;;)
-  {
-    vidTaskBtnProcess();
-    osDelay(1);
-  }
-  /* USER CODE END vidTaskButton */
-}
-
-/* USER CODE BEGIN Header_vidTaskSwd */
-/**
-* @brief Function implementing the TSK_SWD thread.
-* @param argument: Not used
-* @retval None
-*/
-/* USER CODE END Header_vidTaskSwd */
-void vidTaskSwd(void *argument)
-{
-  /* USER CODE BEGIN vidTaskSwd */
-  /* Infinite loop */
-  for(;;)
-  {
-    vidTaskSwdProcess();
-    osDelay(1);
-  }
-  /* USER CODE END vidTaskSwd */
-}
 
 /**
   * @brief  Period elapsed callback in non blocking mode
